+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/features/MinecartFeature.java
@@ -58,7 +58,7 @@ public class MinecartFeature extends Feature<MinecartConfig> {
         structurePlaceSettings.setRotation(Rotation.getRandom(context.random())).setRotationPivot(halfLengths).setIgnoreEntities(false);
         blockpos$Mutable.set(context.origin());
         BlockPos offset = new BlockPos(-template.get().getSize().getX() / 2, 0, -template.get().getSize().getZ() / 2);
-        template.get().placeInWorld(context.level(), blockpos$Mutable.offset(offset), blockpos$Mutable.offset(offset), structurePlaceSettings, context.random(), Block.UPDATE_INVISIBLE);
+        template.get().placeInWorld(context.level(), blockpos$Mutable.offset(offset), blockpos$Mutable.offset(offset), structurePlaceSettings, context.random(), Block.UPDATE_CLIENTS);
 
         return true;
     }